refactor(ui): extract requireUnlockedTab shared route guard
SonarCloud flagged the new /dashboard/security/$dbId route at 61.8%
duplication on PR #245 — its beforeLoad body is character-for-character
identical to the dashboard index route's. Both bodies collapse to a
single requireUnlockedTab(params.dbId) call against the new helper at
src/lib/require-unlocked-tab.ts.
