+++ src/routes/dashboard/index.$dbId.tsx
@@ -1,9 +1,9 @@
-import { createFileRoute, redirect } from "@tanstack/react-router";
+import { createFileRoute } from "@tanstack/react-router";
 import { useIsMobile } from "@/hooks/use-mobile";
 import MobileContentArea from "@/views/MobileContentArea.tsx";
 import DesktopContentArea from "@/views/DesktopContentArea.tsx";
 import { database, KeepassIdSchema } from "@/lib/tauri";
-import { useDatabaseTabs } from "@/stores/database-tabs";
+import { requireUnlockedTab } from "@/lib/require-unlocked-tab";
 import { z } from "zod/v4";
 import { EntrySortFieldSchema, SortOrderSchema } from "@/lib/types";
 
@@ -15,27 +15,7 @@ const DashboardSearchSchema = z.object({
 });
 
 export const Route = createFileRoute("/dashboard/index/$dbId")({
-  beforeLoad: ({ params }) => {
-    const state = useDatabaseTabs.getState();
-    const tab = state.tabs.find(
-      (item) => item.dbId === params.dbId || item.path === params.dbId
-    );
-
-    if (!tab) {
-      throw redirect({ to: "/" });
-    }
-
-    if (tab.state === "unlocking" || tab.state === "locked") {
-      throw redirect({
-        to: "/unlock",
-        search: tab.path ? { path: tab.path } : {},
-      });
-    }
-
-    state.setActiveTab(tab.id);
-
-    return { tabId: tab.id };
-  },
+  beforeLoad: ({ params }) => requireUnlockedTab(params.dbId),
   loaderDeps: ({ search }) => ({ groupId: search.groupId ?? null }),
   loader: async ({ params }) => {
     const info = await database.getInfo(params.dbId);
